main.py
import asyncio
import threading
from typing import Optional
import yaml

from src.camera.camera_input import CameraInput
from src.hand_tracking.hand_detector import HandDetector
from src.t2i_backend.stable_diffusion_model import StableDiffusionModel
from src.prompt_system.spatial_prompts import SpatialPromptManager
from src.ui.interface import MainInterface
from src.tests.prompt_blend_test import PromptBlendTest

class HandT2IApplication:
    """Main application controller"""
    
    def __init__(self, config_path: str = "config/settings.yaml"):
        """
        Initialize application
        
        Args:
            config_path: Path to configuration file
        """
        self.config = self._load_config(config_path)
        self.running = False
        
        # Initialize components
        self.camera: Optional[CameraInput] = None
        self.hand_detector: Optional[HandDetector] = None
        self.t2i_model: Optional[StableDiffusionModel] = None
        self.prompt_manager: Optional[SpatialPromptManager] = None
        self.ui: Optional[MainInterface] = None
        
        # Test modes
        self.current_test = None
        self.tests = {}
    
    def _load_config(self, config_path: str) -> dict:
        """Load configuration from YAML file"""
        # TODO: Implement config loading
        return {}
    
    def initialize_components(self) -> None:
        """Initialize all application components"""
        # TODO: Initialize all components
        pass
    
    def setup_tests(self) -> None:
        """Setup available test modes"""
        # TODO: Setup test instances
        pass
    
    def main_loop(self) -> None:
        """Main application processing loop"""
        # TODO: Implement main loop
        pass
    
    def run(self) -> None:
        """Start the application"""
        self.running = True
        self.initialize_components()
        self.setup_tests()
        
        # Start processing thread
        processing_thread = threading.Thread(target=self.main_loop)
        processing_thread.start()
        
        # Run UI
        self.ui.run()
        
        # Cleanup
        self.running = False
        processing_thread.join()
    
    def shutdown(self) -> None:
        """Gracefully shutdown application"""
        self.running = False
        # TODO: Implement cleanup
        pass

def main():
    """Application entry point"""
    app = HandT2IApplication()
    try:
        app.run()
    except KeyboardInterrupt:
        print("Shutting down...")
        app.shutdown()


if __name__ == "__main__":
    main()