requirements.txt
# Core dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pillow>=10.0.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# UI
customtkinter>=5.2.0

# Utilities
pyyaml>=6.0
typing-extensions>=4.8.0

# Optional for development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0